+++ src/main/java/com/inq/wishhair/wesharewishhair/user/service/MyPageService.java
@@ -4,7 +4,6 @@ import com.inq.wishhair.wesharewishhair.review.service.ReviewFindService;
 import com.inq.wishhair.wesharewishhair.review.service.dto.response.ReviewResponse;
 import com.inq.wishhair.wesharewishhair.user.service.dto.response.MyPageResponse;
 import com.inq.wishhair.wesharewishhair.user.domain.User;
-import com.inq.wishhair.wesharewishhair.user.domain.point.PointHistory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -17,7 +16,7 @@ import java.util.List;
 @Transactional(readOnly = true)
 public class MyPageService {
 
-    private final PointFindService userPointService;
+    private final PointService userPointService;
     private final UserFindService userFindService;
     private final ReviewFindService reviewFindService;
