+++ src/main/java/com/inq/wishhair/wesharewishhair/user/service/PointService.java
@@ -15,7 +15,7 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class PointFindService {
+public class PointService {
 
     private final PointRepository pointRepository;
